# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_kietle_zenreply_pkg_response.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  github_com_kietle_zenreply_pkg_response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kietle_zenreply_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  handler.SlackAuthCallbackResponse:
    properties:
      avatar:
        example: https://avatars.slack-edge.com/...
        type: string
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      slack_id:
        example: U01234567
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handler.SlackAuthURLResponse:
    properties:
      state:
        example: abc123
        type: string
      url:
        example: https://slack.com/oauth/v2/authorize?...
        type: string
    type: object
...
  model.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      slack_name:
        type: string
      slack_team_id:
        type: string
      slack_user_id:
        type: string
      token_scope:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zenreply.app
    name: ZenReply Support
    url: https://zenreply.app/support
  description: ZenReply is an intelligent Slack auto-reply system for deep work sessions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zenreply.app/terms
  title: ZenReply API
  version: 1.0.0
paths:
...
  /slack/auth:
    get:
      description: Returns the Slack OAuth authorization URL for user login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SlackAuthURLResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      summary: Get Slack OAuth URL
      tags:
      - auth
  /slack/callback:
    get:
      description: Exchanges the OAuth code for tokens, upserts the user, and returns
        a JWT
      parameters:
      - description: OAuth authorization code from Slack
        in: query
        name: code
        required: true
        type: string
      - description: State token for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SlackAuthCallbackResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      summary: Handle Slack OAuth callback
      tags:
      - auth
  /slack/callback/redirect:
    get:
      description: Same as /slack/callback but redirects to the frontend with the
        token as a query param
      parameters:
      - description: OAuth authorization code from Slack
        in: query
        name: code
        required: true
        type: string
      - description: State token for CSRF protection
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      summary: Handle Slack OAuth callback with redirect
      tags:
      - auth
  /slack/events:
    post:
      consumes:
      - application/json
      description: Receives and processes events from the Slack Events API (webhook
        mode). Handles URL verification and message events.
      parameters:
      - description: Slack request signature
        in: header
        name: X-Slack-Signature
        required: true
        type: string
      - description: Slack request timestamp
        in: header
        name: X-Slack-Request-Timestamp
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      summary: Slack Events API webhook
      tags:
      - slack
  /users/me:
    delete:
      description: Deactivates the authenticated user's account and revokes access
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Deactivate current user account
      tags:
      - users
    get:
      description: Returns the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your-jwt-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"